# backend/src/main/java/com/dresscode/dto/auth/LoginResponseDto.java
package com.dresscode.dto.auth;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Setter
@Getter
@AllArgsConstructor
@NoArgsConstructor

public class LoginResponseDto {
    private String token;
}